Implement DocfxProject.Build
