+++ samples/plugins/.gitignore
