+++ samples/plugins/src/ExampleClass.cs
