+++ src/Microsoft.DocAsCode.App/DocfxProject.cs
@@ -1,13 +1,11 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.IO;
 using System.Threading.Tasks;
-using Microsoft.DocAsCode.Build.Engine;
 using Microsoft.DocAsCode.Common;
-using Microsoft.DocAsCode.Metadata.ManagedReference;
-using Microsoft.DocAsCode.Plugins;
+using Microsoft.DocAsCode.SubCommands;
+using Newtonsoft.Json.Linq;
 
 namespace Microsoft.DocAsCode
 {
@@ -33,6 +31,16 @@ namespace Microsoft.DocAsCode
 
         public Task Build()
         {
+            var config = JObject.Parse(File.ReadAllText(_configPath));
+            if (config.TryGetValue("metadata", out var value))
+                RunMetadata.Exec(value.ToObject<MetadataJsonConfig>(JsonUtility.DefaultSerializer.Value), Path.GetDirectoryName(_configPath));
+            if (config.TryGetValue("merge", out value))
+                RunMerge.Exec(value.ToObject<MergeJsonConfig>(JsonUtility.DefaultSerializer.Value));
+            if (config.TryGetValue("pdf", out value))
+                RunPdf.Exec(value.ToObject<PdfJsonConfig>(JsonUtility.DefaultSerializer.Value));
+            if (config.TryGetValue("build", out value))
+                RunBuild.Exec(value.ToObject<BuildJsonConfig>(JsonUtility.DefaultSerializer.Value));
+
             return Task.CompletedTask;
         }
     }
