+++ src/docfx/SubCommands/PdfCommand.cs
@@ -1,20 +1,15 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
+using System.IO;
+
+using Newtonsoft.Json;
+using Microsoft.DocAsCode.Common;
+using Microsoft.DocAsCode.Plugins;
+
 namespace Microsoft.DocAsCode.SubCommands
 {
-    using System;
-    using System.Collections.Generic;
-    using System.IO;
-
-    using Newtonsoft.Json;
-
-    using Microsoft.DocAsCode;
-    using Microsoft.DocAsCode.Common;
-    using Microsoft.DocAsCode.Exceptions;
-    using Microsoft.DocAsCode.HtmlToPdf;
-    using Microsoft.DocAsCode.Plugins;
-
     internal sealed class PdfCommand : ISubCommand
     {
         private readonly PdfJsonConfig _config;
@@ -34,70 +29,7 @@ namespace Microsoft.DocAsCode.SubCommands
 
         public static void Exec(PdfJsonConfig config)
         {
-            var wkhtmltopdfFilePath = config.Wkhtmltopdf?.FilePath.GetFullFilePath(config.BaseDirectory);
-            ConvertWrapper.PrerequisiteCheck(wkhtmltopdfFilePath);
-
-            if (config.Serve == true)
-            {
-                Logger.LogWarning("--serve is not supported in pdf command, ignored");
-                config.Serve = false;
-            }
-
-            if (config.Templates == null || config.Templates.Count == 0)
-            {
-                config.Templates = new ListWithStringFallback(new List<string> { "pdf.default" });
-            }
-
-            EnvironmentContext.SetBaseDirectory(Path.GetFullPath(string.IsNullOrEmpty(config.BaseDirectory) ? Directory.GetCurrentDirectory() : config.BaseDirectory));
-            // TODO: remove BaseDirectory from Config, it may cause potential issue when abused
-            var baseDirectory = EnvironmentContext.BaseDirectory;
-            var outputFolder = Path.GetFullPath(Path.Combine(string.IsNullOrEmpty(config.OutputFolder) ? baseDirectory : config.OutputFolder, config.Destination ?? string.Empty));
-            var rawOutputFolder = string.IsNullOrEmpty(config.RawOutputFolder) ? Path.Combine(outputFolder, "_raw") : config.RawOutputFolder;
-            var options = new PdfOptions
-            {
-                BasePath = config.BasePath,
-                CssFilePath = config.CssFilePath,
-                DestDirectory = outputFolder,
-                Host = config.Host,
-                Locale = config.Locale,
-                NeedGeneratePdfExternalLink = config.GeneratesExternalLink,
-                GenerateAppendices = config.GeneratesAppendices,
-                PdfConvertParallelism = config.MaxParallelism == null || config.MaxParallelism <= 0 ? Environment.ProcessorCount : config.MaxParallelism.Value,
-                PdfDocsetName = config.Name ?? Path.GetFileName(EnvironmentContext.BaseDirectory),
-                SourceDirectory = Path.Combine(rawOutputFolder, config.Destination ?? string.Empty),
-                ExcludeTocs = config.ExcludedTocs?.ToArray(),
-                KeepRawFiles = config.KeepRawFiles,
-                ExcludeDefaultToc = config.ExcludeDefaultToc,
-                LoadErrorHandling = config.LoadErrorHandling,
-                FilePath = wkhtmltopdfFilePath,
-                AdditionalPdfCommandArgs = config.Wkhtmltopdf?.AdditionalArguments,
-                TocTitle = config.TocTitle,
-                OutlineOption = config.OutlineOption,
-                CoverPageTitle = config.CoverPageTitle,
-                NoInputStreamArgs = config.NoInputStreamArgs,
-            };
-
-            // 1. call BuildCommand to generate html files first
-            // Output build command exec result to temp folder
-            config.OutputFolder = rawOutputFolder;
-            RunBuild.Exec(config);
-
-            // 2. call html2pdf converter
-            var converter = new ConvertWrapper(options);
-            try
-            {
-                using (new LoggerPhaseScope("PDF", LogLevel.Info))
-                {
-                    Logger.LogInfo("Start generating PDF files...");
-                    converter.Convert();
-                }
-            }
-            catch (IOException ioe)
-            {
-                throw new DocfxException(ioe.Message, ioe);
-            }
-
-            // 3. Should we delete generated files according to manifest
+            RunPdf.Exec(config);
         }
 
         private static PdfJsonConfig ParseOptions(PdfCommandOptions options)
@@ -204,7 +136,7 @@ namespace Microsoft.DocAsCode.SubCommands
 
             if (!string.IsNullOrEmpty(options.FilePath))
             {
-                config.Wkhtmltopdf.FilePath = options.FilePath.GetFullFilePath(Environment.CurrentDirectory);
+                config.Wkhtmltopdf.FilePath = Path.Combine(Environment.CurrentDirectory, options.FilePath);
             }
         }
